enrollment module

+++ src/users/students/students.service.ts
@@ -1,4 +1,9 @@
-import { Inject, Injectable, forwardRef } from '@nestjs/common';
+import {
+  Inject,
+  Injectable,
+  NotFoundException,
+  forwardRef,
+} from '@nestjs/common';
 import { EnrollStudentDTO } from './dto/enroll-student.dto';
 import { Repository } from 'typeorm';
 import { Student } from './entities/student.entity';
@@ -6,6 +11,7 @@ import { CoursesService } from 'src/courses/courses.service';
 import { InjectRepository } from '@nestjs/typeorm';
 import { CreateStudentDto } from './dto/create-student.dto';
 import { UpdateStudentDto } from './dto/update-student.dto';
+import { TasksService } from 'src/tasks/tasks.service';
 
 @Injectable()
 export class StudentsService {
@@ -13,37 +19,62 @@ export class StudentsService {
     @InjectRepository(Student)
     private readonly studentRepository: Repository<Student>,
     @Inject(forwardRef(() => CoursesService))
-    private readonly couseService: CoursesService,
+    private readonly courseService: CoursesService,
+    @Inject(forwardRef(() => TasksService))
+    private readonly taskService: TasksService,
   ) {}
 
   async enrollStudent(enrollStudentDTO: EnrollStudentDTO) {
     const { courseId, studentId } = enrollStudentDTO;
     const [student, course] = await Promise.all([
       this.studentRepository.findOne({ where: { id: studentId } }),
-      this.couseService.findOne(courseId),
+      this.courseService.findOne(courseId),
     ]);
     if (student && course) {
-      
+      if (!student.enrollments) {
+        student.enrollments = [];
+      } else {
+        student.enrollments.push();
+      }
+      this.studentRepository.save(student);
     }
   }
-  
-  create(createAdminDto: CreateStudentDto) {
-    return 'This action adds a new admin';
+
+  async create(createStudentDto: CreateStudentDto) {
+    const { email, password } = createStudentDto;
+    return this.studentRepository.save(new Student(email, password));
   }
 
   async findAll() {
     return this.studentRepository.find();
   }
 
-  findOne(id: string) {
-    return this.studentRepository.findOne({where:{id: id}});
+  async findOne(id: string) {
+    return this.studentRepository.findOne({ where: { id: id } });
   }
 
-  update(id: string, updateAdminDto: UpdateStudentDto) {
-    return `This action updates a #${id} admin`;
+  async update(id: string, updateStudentDto: UpdateStudentDto) {
+    return `This action updates a #${id} student`;
   }
 
-  remove(id: string) {
+  async remove(id: string) {
+    const student = await this.studentRepository.findOne({
+      where: {
+        id: id,
+      },
+      relations: ['courses', 'tasks'],
+    });
+
+    if (!student) throw new NotFoundException('Student not found');
+
+    // await Promise.all(
+    //   student.courses.map((course) => this.courseService.remove(course.id)),
+    // );
+
+    await Promise.all(
+      student.tasks.map((task) => this.taskService.remove(task.id)),
+    );
+
     return this.studentRepository.delete(id);
   }
 }
